+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add use System dialog feature
 - Solve the page jump scene on the dialog (useSystem)
 - SmartStatus add status：custom，attach，allCustom
+- The entry 'init' method can customize the default style of Toast and Loading
 - Fix bug
 
 # [3.2.x]